update test structure and renamed test classes so they get picked up by dart test
