+++ lib/src/utils/extensions.dart
@@ -21,6 +21,8 @@ extension StringListExtension on List<String> {
   }
 }
 
+
+
 extension Uint8ListExtension on Uint8List {
 
   /// Converts a [Uint8List] to a hex string.
