+++ test/crypto/hash/blake2b_test.dart
@@ -42,12 +42,9 @@ main() {
 
       test('throws error when bytes is empty', () {
         final blake2b = Blake2b256();
-        // expect(() => blake2b.hash([]), throwsA(isA<AssertionError>()));
 
-        print(blake2b.hash(Uint8List(0)).toHexString());
         expect(blake2b.hash(Uint8List(0)).toHexString(),
             equals("0e5751c026e543b2e8ab2eb06099daa1d1e5df47778f7787faab45cdf12fe3a8"));
-        // expect(() => blake2b.hash([]), equals("0e5751c026e543b2e8ab2eb06099daa1d1e5df47778f7787faab45cdf12fe3a8"));
       });
     });
